rename BullyText2015ABTest.js to PlayerNamesInGame.js, add new field to competitiveStoriesConfigModel.js
